refactor: move legacy client module into clients
