+++ src/ai/backend/manager/clients/container_registry/harbor.py
