+++ src/ai/backend/manager/services/artifact_revision/service.py
@@ -26,7 +26,7 @@ from ai.backend.common.dto.storage.request import (
     ReservoirImportModelsReq,
 )
 from ai.backend.logging import BraceStyleAdapter
-from ai.backend.manager.client.artifact_registry.reservoir_client import ReservoirRegistryClient
+from ai.backend.manager.clients.artifact_registry.reservoir_client import ReservoirRegistryClient
 from ai.backend.manager.clients.storage_proxy.session_manager import StorageSessionManager
 from ai.backend.manager.config.provider import ManagerConfigProvider
 from ai.backend.manager.config.unified import (
